clisqlclient: drop the dependency on 'security'
The 'security' package is quite large. There is no reason to embark it
into a standalone SQL shell.

Towards achieving this, this patch removes the dependency on the
'security' package from `clisqlclient` by moving the password prompt
code into a sub-package.

Release note: None
