+++ pkg/security/BUILD.bazel
@@ -44,7 +44,6 @@ go_library(
         "@org_golang_x_crypto//bcrypt",
         "@org_golang_x_crypto//ocsp",
         "@org_golang_x_sync//errgroup",
-        "@org_golang_x_term//:term",
     ],
 )
